+++ components/layout/SidebarMain.tsx
@@ -34,6 +34,96 @@ import { Command, CommandInput, CommandItem, CommandList } from '@/components/ui
 import { Dialog, DialogClose, DialogContent, DialogTrigger } from '@/components/ui/dialog';
 import { Conversation, useConversations } from '@/components/interactive/useConversation';
 import { InteractiveConfigContext } from '@/components/interactive/InteractiveConfigContext';
+import { useCompany } from '@/components/interactive/useUser';
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from '@/components/ui/dropdown-menu';
+import { setCookie } from 'cookies-next';
+import { Check, ChevronsUpDown, Plus } from 'lucide-react';
+import { FaRobot } from 'react-icons/fa';
+import { Agent, useAgent, useAgents } from '@/components/interactive/useAgent';
+
+export function AgentSelector() {
+  const { isMobile } = useSidebar('left');
+  const { data: activeAgent, mutate: mutateActiveAgent, error: agentError } = useAgent();
+  const { data: activeCompany, mutate: mutateActiveCompany, error: companyError } = useCompany();
+  const { data: agentsData } = useAgents();
+  const router = useRouter();
+  console.error({ agentError, companyError });
+
+  const switchAgents = (agent: Agent) => {
+    // setActiveAgent(agent);
+    setCookie('agixt-agent', agent.name, {
+      domain: process.env.NEXT_PUBLIC_COOKIE_DOMAIN,
+    });
+    mutateActiveCompany();
+    mutateActiveAgent();
+  };
+
+  return (
+    <SidebarMenu>
+      <SidebarMenuItem>
+        <DropdownMenu>
+          <DropdownMenuTrigger asChild>
+            <SidebarMenuButton
+              side='left'
+              size='lg'
+              className='data-[state=open]:bg-sidebar-accent data-[state=open]:text-sidebar-accent-foreground'
+            >
+              <div className='flex items-center justify-center rounded-lg aspect-square size-8 bg-sidebar-primary text-sidebar-primary-foreground'>
+                <FaRobot className='size-4' />
+              </div>
+              <div className='grid flex-1 text-sm leading-tight text-left'>
+                <span className='font-semibold truncate'>{activeAgent?.agent?.name}</span>
+                <span className='text-xs truncate'>{activeCompany?.name}</span>
+              </div>
+              <ChevronsUpDown className='ml-auto' />
+            </SidebarMenuButton>
+          </DropdownMenuTrigger>
+          <DropdownMenuContent
+            className='w-[--radix-dropdown-menu-trigger-width] min-w-64 rounded-lg px-2'
+            align='start'
+            side={isMobile ? 'bottom' : 'right'}
+            sideOffset={4}
+          >
+            <DropdownMenuLabel className='text-xs text-muted-foreground'>Agents</DropdownMenuLabel>
+            {agentsData &&
+              agentsData.map((agent) => (
+                <DropdownMenuItem
+                  key={agent.id}
+                  onClick={() => switchAgents(agent)}
+                  className='flex items-center justify-between p-2 cursor-pointer'
+                >
+                  <div className='flex flex-col'>
+                    <span>{agent.name}</span>
+                    <span className='text-xs text-muted-foreground'>{agent.companyName}</span>
+                  </div>
+                  {activeAgent?.agent?.id === agent.id && <Check className='w-4 h-4 ml-2' />}
+                </DropdownMenuItem>
+              ))}
+            <DropdownMenuSeparator />
+            <DropdownMenuItem
+              className='gap-2 p-2 cursor-pointer'
+              onClick={() => {
+                router.push('/settings');
+              }}
+            >
+              <div className='flex items-center justify-center border rounded-md size-6 bg-background'>
+                <Plus className='size-4' />
+              </div>
+              <div className='font-medium text-muted-foreground'>Add Agent</div>
+            </DropdownMenuItem>
+          </DropdownMenuContent>
+        </DropdownMenu>
+      </SidebarMenuItem>
+    </SidebarMenu>
+  );
+}
 
 export function ChatHistory() {
   const state = useContext(InteractiveConfigContext);
